base for testing appears to be up
