+++ fastAPI/test_main.py
@@ -1,24 +1,41 @@
 import pytest
 from fastapi.testclient import TestClient
-from main import app, get_test_db
+from main import app, get_db
+from models import Base, User, GenrePref
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+# test_main.py
+from database import test_engine, get_test_db  # Assuming your test_engine is defined in database.py
+
+
+# Test Database Setup
+TEST_DATABASE_URL = "sqlite:///:memory:"  # In-memory SQLite database
+test_engine = create_engine(TEST_DATABASE_URL, connect_args={"check_same_thread": False})
+TestingSessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=test_engine)
+
+
+@pytest.fixture(scope="module", autouse=True)
+def setup_test_database():
+    # Create tables in the in-memory SQLite database
+    Base.metadata.create_all(bind=test_engine)
+
+    # Override the app's get_db dependency
+    app.dependency_overrides[get_db] = lambda: get_test_db()
+
+    # Yield to the test
+    yield
+
+    # Teardown: Drop tables after tests
+    Base.metadata.drop_all(bind=test_engine)
 
 # Use TestClient to simulate API requests
 client = TestClient(app)
 
-
-@pytest.fixture
-def client():
-    app.dependency_overrides[get_db] = get_test_db
-    with TestClient(app) as client:
-        yield client
-
-# Example test for a POST endpoint
+# Example test for POST /user/{username}/genrepref
 def test_post_genrepref():
-    username = "testuser"
-    genre_id = 1
-    response = client.post(f"/user/{username}/genrepref", params={"genreID": genre_id})
-    assert response.status_code == 200 or response.status_code == 400
-    if response.status_code == 200:
-        assert response.json()["message"] == "Genre added"
-    elif response.status_code == 400:
-        assert response.json()["detail"] == "Game is already in the genrePref"
+    # Arrange: Add a test user
+    test_username = "testuser"
+    with next(get_test_db()) as db:  # Extract session from the generator
+        db.add(User(username=test_username))
+        db.commit()
+
